requirements.txt
flask
flask-cors
gunicorn
tensorflow
numpy
pillow
